+++ release/ReleaseNotes.md
@@ -8,6 +8,7 @@
 
 - Removed **EnhancementParameter::NoiseGateEnable** as it is now a fixed part of our VAD.
 - Added new error code **ErrorCode::ParameterFixed** returned when attempting to modify a parameter of a model with fixed parameters.
+- Simplified `AicVad::is_speech_detected()` to return `bool` directly instead of using an output parameter with error code.
 
 ### Fixes